face_recognition/requirements.txt
opencv-python==4.8.1
numpy==1.24.3
face-recognition==1.3.0
dlib==19.24.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
torch==2.1.2
torchvision==0.16.2
pillow==10.1.0 